refactor: move load county
